Fluids are now more async safe

+++ src/main/java/ac/grim/grimac/checks/movement/predictions/PredictionEngineNormal.java
@@ -15,7 +15,7 @@ public class PredictionEngineNormal extends PredictionEngine {
             vector.setZ(Mth.clamp(vector.getZ(), -0.15, 0.15));
             vector.setY(Math.max(vector.getY(), -0.15));
 
-            if (vector.getY() < 0.0 && !(ChunkCache.getBlockDataAt((int) grimPlayer.lastX, (int) grimPlayer.lastY, (int) grimPlayer.lastZ).getBlock() instanceof BlockScaffolding) && grimPlayer.isSneaking && !grimPlayer.isFlying) {
+            if (vector.getY() < 0.0 && !(ChunkCache.getBlockDataAt(grimPlayer.lastX, grimPlayer.lastY, grimPlayer.lastZ).getBlock() instanceof BlockScaffolding) && grimPlayer.isSneaking && !grimPlayer.isFlying) {
                 vector.setY(0.0);
             }
         }
